Move CLI modules to utils.cli
